requirements.txt
threadpool==1.3.2
websockets==8.1
selenium==3.141.0
PyYAML==5.3.1
pymysql==0.9.3
psycopg2-binary==2.8.6
DBUtils==1.3
requests==2.25.1
openpyxl==3.0.3
xlrd==1.2.0
colorama==0.4.4